refactor: consolidate schema merge into SchemaBuilder with single entry point
- Replace MergeSchemaWithPolicy + convertSchema with unified BuildSchemaConfig(cr, upstream)
- Move all merge logic into SchemaBuilder as private methods (fromCR, mergeWithUpstream, mergeTables, mergeColumns)
- Delete schema_merge.go, remove convertSchema/processSchemaTable from controller utils
- Rename applySchemaWithMerge to applySchema
- Consolidate tests into schema_builder_test.go
